+++ packages/eslint-config-dawn/legacy.js
@@ -1,3 +1,3 @@
 module.exports = {
-  extends: require.resolve('eslint-config-ali/es5'),
+  extends: [require.resolve('eslint-config-ali/es5')],
 };