+++ packages/eslint-config-dawn/ts-react.js
@@ -1,4 +1,4 @@
 // Alias for typescript-react.js
 module.exports = {
-  extends: require.resolve('./typescript-react'),
+  extends: [require.resolve('./typescript-react')],
 };